OCSADV-200: move program copy to RichProgram
